correction

+++ src/main/java/com/orsys/model/Math.java
@@ -31,7 +31,7 @@ public String max()
 
 public String factoriel()
 {
-	int f=1;
+	long f=1;
 	for(int i=2;i<=this.a;i++)
 		f=f*i;
 	return "la factoriel est"+ f;
